requirements.txt
elephant==0.10.0
mat4py==0.5.0
matplotlib==3.5.0
neo==0.11.0
numpy==1.22.4
pandas==1.3.4
PyYAML==6.0
QtPy==1.11.3
quantities==0.12.5
scikit-learn==1.0.1
scipy==1.7.3
sklearn==0.0
